+++ pyroomacoustics/bss/__init__.py
@@ -25,26 +25,27 @@ A few commonly used functions, such as projection back, can be found in
 References
 ----------
 
-.. [1] N. Ono, *Stable and fast update rules for independent vector analysis
-    based on auxiliary function technique*, Proc. IEEE, WASPAA, 2011.
+.. [1] N. Ono, *Stable and fast update rules for independent vector analysis based
+    on auxiliary function technique,* Proc. IEEE, WASPAA, pp. 189-192, Oct. 2011.
+
 
 .. [2] R. Aichner, H. Buchner, F. Yan, and W. Kellermann  *A real-time
     blind source separation scheme and its application to reverberant and noisy
     acoustic environments*,  Signal Processing, 86(6), 1260-1277.
     doi:10.1016/j.sigpro.2005.06.022, 2006.
 
-.. [3] D. Kitamura, N. Ono, H. Sawada, H. Kameoka, and H. Saruwatari
-    *Determined Blind Source Separation with Independent Low-Rank Matrix Analysis*,
-    in Audio Source Separation, S. Makino, Ed. Springer, 2018, pp.  125-156.
+.. [3] D. Kitamura, N. Ono, H. Sawada, H. Kameoka, H. Saruwatari, *Determined blind
+    source separation unifying independent vector analysis and nonnegative matrix
+    factorization,* IEEE/ACM Trans. ASLP, vol. 24, no. 9, pp. 1626-1641, Sept. 2016
 
-.. [4] Janský, Jakub & Koldovský, Zbyněk & Ono, Nobutaka. (2016). A computationally
-    cheaper method for blind speech separation based on AuxIVA and incomplete demixing
-    transform. 1-5. 10.1109/IWAENC.2016.7602921.
+.. [4] J. Janský, Z. Koldovský, and N. Ono, *A computationally cheaper method
+    for blind speech separation based on AuxIVA and incomplete demixing transform,*
+    Proc. IEEE, IWAENC, pp. 1-5, Sept. 2016.
 
 '''
 
 from .trinicon import trinicon
-from .auxiva import auxiva, f_contrasts
+from .auxiva import auxiva
 from .ilrma import ilrma
 from .sparseauxiva import sparseauxiva
 from .common import projection_back, sparir